Password Logic
Large add here is some password logic. If the PIN/Pass is empty it does not break the entire script. I also added some logic to tell you if the password is a good password.

There is a new bug with the Decryption right now. It seems to work, but still does not do the right things. Need to look into that more.